+++ DreamCleaningBackend/Models/ServiceType.cs
@@ -16,7 +16,7 @@ namespace DreamCleaningBackend.Models
 
         [StringLength(500)]
         public string? Description { get; set; }
-        public int TimeDuration { get; set; } = 90;
+        public decimal TimeDuration { get; set; } = 90;
         public bool IsActive { get; set; } = true;
         public int DisplayOrder { get; set; }
         public bool HasPoll { get; set; } = false;
